修正启动器配置

+++ base-remove.sh
@@ -37,9 +37,14 @@ rm -rf /etc/systemd/system/systemd-nspawn@$1.service.d
 rm -f /etc/systemd/nspawn/$1.nspawn
 [[ -d /etc/systemd/nspawn && `ls -A /etc/systemd/nspawn |wc -w` == 0 ]] && rm -rf /home/share /etc/systemd/nspawn
 
-rm -f /usr/share/pixmaps/com.qq.im.deepin.svg
-rm -f /usr/share/pixmaps/com.qq.weixin.deepin.svg
-rm -f /usr/share/applications/deepin-qq.desktop
-rm -f /usr/share/applications/deepin-weixin.desktop
+if [ -f /usr/share/applications/deepin-qq.desktop ] && [[ $(cat /usr/share/applications/deepin-qq.desktop | grep $1-) ]]; then
+    rm -f /usr/share/pixmaps/com.qq.im.deepin.svg
+    rm -f /usr/share/applications/deepin-qq.desktop
+fi
+
+if [ -f /usr/share/applications/deepin-weixin.desktop ] && [[ $(cat /usr/share/applications/deepin-weixin.desktop | grep $1-) ]]; then
+    rm -f /usr/share/pixmaps/com.qq.weixin.deepin.svg
+    rm -f /usr/share/applications/deepin-weixin.desktop
+fi
 
 [[ ! $EXEC_FROM_CONFIG ]] && echo "为防止数据意外丢失，您需要手动删除 ~/.machines/$1 文件夹！"
